ms-inscripcion-clase se agrega verificacion en service delete

+++ ms-inscripcion-clase/src/main/java/com/example/ms_inscripcion_clase/service/InscripcionClaseService.java
@@ -65,6 +65,9 @@ public class InscripcionClaseService {
     }
 
     public void delete(Long id){
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("No se puede eliminar inscripcion no encontrada");
+        }
         repository.deleteById(id);
 
     }
